+++ assets/javascripts/kitten/components/form/checkbox.js
@@ -9,13 +9,20 @@ export class Checkbox extends React.Component {
             children,
             text,
             inputClassName,
+            error,
             ...inputProps } = this.props
+
+    const checkboxInputClassNames = classNames(
+      'k-Checkbox__input',
+      inputClassName,
+      { 'is-error': error },
+    )
 
     return (
       <div className={ classNames('k-Checkbox', className) }>
         <input id={ id }
                type="checkbox"
-               className={ classNames('k-Checkbox__input', inputClassName) }
+               className={ checkboxInputClassNames }
                { ...inputProps } />
 
         <label htmlFor={ id } className="k-Checkbox__label">