+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Fix wrong padding for `InformationBox` component.
+
 ## [8.0.0] - 2017-02-28
 
 Breaking change: