+++ CHANGELOG.md
@@ -7,9 +7,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add new `Alert` component.
 - Feature: Add new `CloseButton` component.
 - Feature: Add new `CrossIcon` component.
+- Feature: Add new `error` props on `RadioButton` and `Checkbox`.
 - Feature: Add new `LocationIcon` React component.
 - Feature: Add new `LocationInput` React component.
 - Fix: Fix wrong padding for `InformationBox` component.
+- Fix: Use `handleChange` handler instead of `handleInput` in
+  `TextInputWithLimit` to prevent React warning.
 
 ## [8.0.0] - 2017-02-28