+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add new `error` props on `RadioButton` and `Checkbox`.
 - Feature: Add new `LocationIcon` React component.
 - Feature: Add new `LocationInput` React component.
+- Fix: Use `CloseButton` component instead of simple button in `Popover`.
 - Fix: Fix wrong padding for `InformationBox` component.
 - Fix: Use `handleChange` handler instead of `handleInput` in
   `TextInputWithLimit` to prevent React warning.